Improve many things
This commit goes through and improves a lot and fixes a lot of bugs. I lost track of what since it was so many little things. The biggest thing was that it got rid of the onDone model and switched to using NotificationCenter for updates/creations, so that all loaded table views would be updated, not just the direct parent.