Add authentication view
This commit adds the authentication view and sets it up to launch each time the app wasn’t in the foreground. It handles this in the AppDelegate, presenting the authentication view on top of the frontmost view controller each time the app enters the background or comes back to foreground.